review: extract formatting concerns into formatting.dart
The in-process formatter, the ignore-directive helpers, and the
subprocess Formatter all lived as siblings inside file_renderer.dart.
That mixed layers — FileRenderer's job is wiring templates and
visitor walks, not knowing how to read pubspec.lock or compose a
`// ignore_for_file:` directive.

New module lib/src/render/formatting.dart holds:
- longLineIgnoreBlock / commentReferencesIgnoreBlock — the per-file
  directive blocks.
- maybeAddLongLineIgnore / maybeAddCommentReferencesIgnore /
  maybeAddIgnoreDirectives — the gen-time helpers wired into
  _renderSpecTemplate.
- Formatter — the subprocess wrapper that runs dart pub get / format
  / fix --apply at the end of a regen.
- FormatContext / FormatterConfig — the dart_style settings resolved
  from the consuming package's pubspec and analysis_options.yaml.
- DartFileFormatter — the in-process DartFormatter wrapper with the
  cached resolution. Used by FileRenderer at every .dart write.

FileRenderer now holds a DartFileFormatter field, calls into it
from _renderTemplate / _renderSpecTemplate, and re-exports Formatter
+ RunProcess for the existing FileRendererConfig contract.
The previously-private _FormatContext / _FormatterConfig become
public so cross-file callers (and tests) can name them.

Tests for the helpers move to test/render/formatting_test.dart
alongside ~12 new DartFileFormatter tests covering: pass-through on
non-Dart paths, formatting Dart content, pubspec.lock vs pubspec.yaml
priority, formatter: page_width / trailing_commas read from
analysis_options.yaml, relative include: following (workspace setup),
package: include skip, walk-up to workspace root, default fallbacks,
single resolution caching across many calls.

Verification:
- dart analyze clean.
- dart test: 544 pass.
- github regen dart analyze clean.
