+++ test/render/file_renderer_test.dart
@@ -2414,239 +2414,6 @@ void main() {
     });
   });
 
-  group('maybeAddLongLineIgnore', () {
-    test('passes through short-only content', () {
-      final content =
-          '${List.generate(5, (i) => 'var x$i = $i;').join('\n')}\n';
-      expect(maybeAddLongLineIgnore(content), content);
-    });
-
-    test('prepends directive when a non-import line exceeds 80 cols', () {
-      final longLine = 'final x = ${'a' * 80};';
-      final content = 'var x = 1;\n$longLine\n';
-      expect(
-        maybeAddLongLineIgnore(content),
-        startsWith('$longLineIgnoreBlock\n'),
-      );
-    });
-
-    test('is idempotent — does not stack the directive', () {
-      final longLine = 'final x = ${'a' * 80};';
-      final content = '$longLineIgnoreBlock\n$longLine\n';
-      // Directive present exactly once before and after.
-      expect(longLineIgnoreBlock.allMatches(content).length, 1);
-      final result = maybeAddLongLineIgnore(content);
-      expect(longLineIgnoreBlock.allMatches(result).length, 1);
-    });
-
-    test('passes through when only long lines are imports', () {
-      // Mirrors the analyzer's `lines_longer_than_80_chars` carve-out
-      // for URI imports — flagging these would trip `unnecessary_ignore`
-      // because the underlying lint never fires.
-      final longImport =
-          "import 'package:${'a' * 60}/foo.dart' as ${'b' * 30};";
-      final content = '$longImport\nclass Foo {}\n';
-      expect(longImport.length, greaterThan(80));
-      expect(maybeAddLongLineIgnore(content), content);
-    });
-
-    test('passes through when only long lines are exports', () {
-      final longExport =
-          "export 'package:${'a' * 60}/foo.dart' show ${'b' * 30};";
-      final content = '$longExport\nclass Foo {}\n';
-      expect(longExport.length, greaterThan(80));
-      expect(maybeAddLongLineIgnore(content), content);
-    });
-
-    test('fires when a `///` doc comment exceeds 80 cols', () {
-      // The analyzer's lines_longer_than_80_chars lint flags long
-      // doc comments — only URI imports/exports are exempt. Verified
-      // empirically: github regen surfaced 37 such hits before this
-      // carve-out was removed.
-      final longDoc = '/// ${'a' * 90}';
-      final content = '$longDoc\nclass Foo {}\n';
-      expect(longDoc.length, greaterThan(80));
-      expect(
-        maybeAddLongLineIgnore(content),
-        startsWith('$longLineIgnoreBlock\n'),
-      );
-    });
-
-    test('still fires when a real long line sits next to a long import', () {
-      final longImport =
-          "import 'package:${'a' * 60}/foo.dart' as ${'b' * 30};";
-      final longCode = 'final x = ${'a' * 80};';
-      final content = '$longImport\n$longCode\n';
-      expect(
-        maybeAddLongLineIgnore(content),
-        startsWith('$longLineIgnoreBlock\n'),
-      );
-    });
-
-    test('treats exactly 80 chars as not over', () {
-      final exactly80 = 'x' * 80;
-      final content = '$exactly80\n';
-      expect(maybeAddLongLineIgnore(content), content);
-    });
-  });
-
-  group('maybeAddCommentReferencesIgnore', () {
-    test('passes through content with no bracketed dartdoc tokens', () {
-      const content = '/// A class with no bracket refs.\nclass Foo {}\n';
-      expect(maybeAddCommentReferencesIgnore(content), content);
-    });
-
-    test('prepends directive when dartdoc has prose-style placeholder', () {
-      // Mirrors github's `code_of_conduct.dart` ("contacting the project
-      // team at [EMAIL]") and `license.dart` (`[year] [fullname]`).
-      const content =
-          '/// Reach out at [EMAIL] for support.\nclass CodeOfConduct {}\n';
-      expect(
-        maybeAddCommentReferencesIgnore(content),
-        startsWith('$commentReferencesIgnoreBlock\n'),
-      );
-    });
-
-    test('passes through `[Foo](url)` markdown links', () {
-      const content =
-          '/// See [the docs](https://example.com).\nclass Bar {}\n';
-      expect(maybeAddCommentReferencesIgnore(content), content);
-    });
-
-    test('does not match plain-comment brackets, only `///` doc comments', () {
-      // Inline `// [SOMETHING]` comments are invisible to dartdoc and
-      // can't trip `comment_references`.
-      const content = '// [PLACEHOLDER] not a dartdoc.\nclass Foo {}\n';
-      expect(maybeAddCommentReferencesIgnore(content), content);
-    });
-
-    test('passes through when every ref is a same-file class', () {
-      // The most common case for spec-derived model files: the doc
-      // comment refers to the class the file declares. The lint
-      // wouldn't fire on these — no need to suppress.
-      const content =
-          '/// Convert a Map to a [Foo].\n'
-          '/// Round-trips with [Foo.fromJson].\n'
-          'class Foo {}\n';
-      expect(maybeAddCommentReferencesIgnore(content), content);
-    });
-
-    test('passes through same-file enum and extension-type refs', () {
-      // The renderer emits all three top-level declaration kinds
-      // (class, enum, extension type); they all need to count as
-      // in-scope.
-      const content =
-          '/// See [Mood] and [Tag].\n'
-          'enum Mood { happy, sad }\n'
-          'extension type const Tag._(String value) {}\n';
-      expect(maybeAddCommentReferencesIgnore(content), content);
-    });
-
-    test('prepends when even one ref escapes the same-file set', () {
-      // `[Foo]` resolves locally, but `[ImportedThing]` does not
-      // (imports aren't tracked yet — see #142). One unresolved ref
-      // is enough to need the directive.
-      const content =
-          '/// [Foo] uses an [ImportedThing] under the hood.\n'
-          'class Foo {}\n';
-      expect(
-        maybeAddCommentReferencesIgnore(content),
-        startsWith('$commentReferencesIgnoreBlock\n'),
-      );
-    });
-
-    test('passes through refs to fields declared in the same file', () {
-      // Surfaced by shorebird_code_push_protocol's
-      // `create_patch_artifact_request.dart`: `/// The signature of
-      // the [hash].` refers to a `final String hash;` field. The
-      // analyzer resolves this through the enclosing class scope and
-      // doesn't fire `comment_references`. Without tracking field
-      // declarations, the helper would add a directive that
-      // `dart fix --apply` then strips as `unnecessary_ignore`,
-      // leaving an orphaned 5-line justification comment.
-      const content =
-          '/// The signature of the [hash].\n'
-          'class CreatePatchArtifactRequest {\n'
-          '  final String hash;\n'
-          '}\n';
-      expect(maybeAddCommentReferencesIgnore(content), content);
-    });
-
-    test('still prepends for refs that look like fields but are not', () {
-      // MIT license templates use `[year]` and `[fullname]` as
-      // placeholders. They look like field names but no field is
-      // declared with those names, so `comment_references` would
-      // fire and the directive is needed.
-      const content =
-          '/// Copyright (c) [year] [fullname].\n'
-          'class License {}\n';
-      expect(
-        maybeAddCommentReferencesIgnore(content),
-        startsWith('$commentReferencesIgnoreBlock\n'),
-      );
-    });
-
-    test('collects multiple refs on a single doc comment line', () {
-      // Greedy regexes have a habit of seeing only the first or last
-      // bracket on a line. Both refs need to be considered: if any
-      // one fails to resolve, we suppress.
-      const contentAllResolve =
-          '/// Compare [A] with [B] and [C].\n'
-          'class A {}\nclass B {}\nclass C {}\n';
-      expect(
-        maybeAddCommentReferencesIgnore(contentAllResolve),
-        contentAllResolve,
-      );
-
-      const contentMiddleEscapes =
-          '/// Compare [A] with [Missing] and [C].\n'
-          'class A {}\nclass C {}\n';
-      expect(
-        maybeAddCommentReferencesIgnore(contentMiddleEscapes),
-        startsWith('$commentReferencesIgnoreBlock\n'),
-      );
-    });
-
-    test('member refs resolve via the head before the dot', () {
-      // `[Foo.fromJson]` resolves iff `Foo` does — the analyzer walks
-      // the head against scope, then looks up the member on whatever
-      // it finds. Test both the resolves-via-head case and the head-
-      // doesn't-resolve case.
-      const resolves =
-          '/// Round-trips through [Foo.fromJson] and [Foo.toJson].\n'
-          'class Foo {}\n';
-      expect(maybeAddCommentReferencesIgnore(resolves), resolves);
-
-      const headMissing =
-          '/// Calls [Missing.method] internally.\nclass Foo {}\n';
-      expect(
-        maybeAddCommentReferencesIgnore(headMissing),
-        startsWith('$commentReferencesIgnoreBlock\n'),
-      );
-    });
-
-    test('refs are checked against every top-level declaration', () {
-      // A real generated file can declare more than one class (e.g. a
-      // model plus an inner enum). The declaration scan has to find
-      // them all, not just the first one.
-      const content =
-          '/// [Outer] holds a list of [Inner].\n'
-          'class Outer {}\n\nenum Inner { a, b }\n';
-      expect(maybeAddCommentReferencesIgnore(content), content);
-    });
-
-    test('indented `///` member docs still count', () {
-      // Generated classes sit at the top level but their member dartdoc
-      // lives inside the class body, indented. The token scan must not
-      // require the `///` to be at column 0.
-      const content =
-          'class Foo {\n'
-          '  /// Members of [Foo] go here.\n'
-          '  void doThing() {}\n'
-          '}\n';
-      expect(maybeAddCommentReferencesIgnore(content), content);
-    });
-  });
 
   // While we still support logging, this should no longer happen since
   // we detect collisions and fix them during resolution.
